Allow to pass variables when triggering pipelines
The implementation for dealing with the split is very naïve, but for
most scenarios I think it'll work well enough.

That said the long term goal is to make a more robust way of defining
the CLI arguments a command supports, so this is a stop gap measure.
(Now lets see how far into the future I can look back and curse at
this comment because it's still true :p)