[#8] Expand user home path from effective user
Under certain circumstances you might run gocd-cli where the original
user was root, which has done sudo -u <other user> to run as another
user. In those cases the $HOME should be that of the effective user,
the one that sudo is being run for, and not the one who called sudo.

The way this is normally done in Python is by checking the os
environment variable $HOME and replacing ~/paths to effectively be
$HOME/paths.

For instance:

  NRPE is a daemon that gets started as root. It then daemonizes and
  starts running as nagios. Any applications being run through that
  daemon will have the effective user nagios. But with the environment
  of the original user, root, so gocd-cli will think the path to the
  config file is "/root/.gocd" instead of "/home/nagios/.gocd".

This closes #8